Do not fail when deleting a nonexistent object
If the object specified in the request is not found, s3rver responds
with 404 error, unlike Amazon S3 that returns the result as deleted
for both single and multi-object operations.

Align deleteObject and deleteObjects with the API spec and return
a success response even when the target object does not exist.