Decode XML entities in elements in config files
Needed for WebsiteConfiguration where you're doing a
ReplaceKeyPrefixWith to something that puts the Key into a QueryString.

e.g. `/thumbnail/image.png` is transformed to `/resize?size=thumbnail&key=image.png`

The & in the query strings have to be encoded in the website config,
and without this change they stay encoded and break things.